+++ CHANGES.md
@@ -36,6 +36,7 @@ Changes in 12.8
 - Skip CSP violation triggered by Facebook Click Identifier in SaveReport.php
 - Fix SQL error when staff is manually created in next school year in Rollover.php
 - Fix SQL error when school was created after Rollover and not rolling schools in Rollover.php
+- SQL add "Review RosarioSIS" link to Resources in rosariosis.sql & rosariosis_mysql.sql
 
 Changes in 12.7.5
 -----------------